+++ ui/mods/com.pa.quitch.qaipersonalities/new_game.js
@@ -1,22 +1,22 @@
-var qaipersonalitiesLoaded;
+var aipPersonalitiesLoaded;
 
-function qaipersonalities() {
-  if (qaipersonalitiesLoaded) {
+function aipPersonalities() {
+  if (aipPersonalitiesLoaded) {
     return;
   }
 
-  qaipersonalitiesLoaded = true;
+  aipPersonalitiesLoaded = true;
 
-  api.debug.log("Adding AI Personalities for Skirmish and Multiplayer");
+  api.debug.log("Adding AI Personalities for Skirmish and Multiplayer loaded");
 
   var aiPersonalities = model.aiPersonalities();
 
   var newPersonalities = {
-    "AIP Aggressive": {
+    aipAggressive: {
       display_name: "!LOC:AIP Aggressive",
       neural_data_mod: 1.33,
     },
-    "AIP Air": {
+    aipAir: {
       display_name: "!LOC:AIP Air",
       percent_vehicle: 0,
       percent_bot: 0,
@@ -24,7 +24,7 @@ function qaipersonalities() {
       percent_naval: 0,
       percent_orbital: 0.05,
     },
-    "AIP Bot": {
+    aipBot: {
       display_name: "!LOC:AIP Bot",
       percent_vehicle: 0,
       percent_bot: 0.75,
@@ -32,12 +32,12 @@ function qaipersonalities() {
       percent_naval: 0,
       percent_orbital: 0.05,
     },
-    "AIP Cautious": {
+    aipCautious: {
       display_name: "!LOC:AIP Cautious",
       neural_data_mod: 0.75,
       min_basic_fabbers: 4,
     },
-    "AIP Defender": {
+    aipDefender: {
       display_name: "!LOC:AIP Defender",
       metal_drain_check: 0.71,
       energy_drain_check: 0.8,
@@ -46,7 +46,7 @@ function qaipersonalities() {
       adv_eco_mod: 1,
       min_basic_fabbers: 4,
     },
-    "AIP Economist": {
+    aipEconomist: {
       display_name: "!LOC:AIP Economist",
       percent_open_vehicle: 0,
       percent_open_bot: 1,
@@ -58,7 +58,7 @@ function qaipersonalities() {
       adv_eco_mod: 1,
       min_basic_fabbers: 4,
     },
-    "AIP Fast Tech": {
+    aipFastTech: {
       display_name: "!LOC:AIP Fast Tech",
       percent_open_vehicle: 0.5,
       percent_open_bot: 0.5,
@@ -69,7 +69,7 @@ function qaipersonalities() {
       adv_eco_mod_alone: 0.5,
       min_basic_fabbers: 4,
     },
-    "AIP Foundation": {
+    aipFoundation: {
       display_name: "!LOC:AIP Foundation",
       fabber_to_factory_ratio_basic: 1.5,
       min_basic_fabbers: 1,
@@ -84,7 +84,7 @@ function qaipersonalities() {
       metal_demand_check: 0.75,
       energy_demand_check: 0.85,
     },
-    "AIP Land": {
+    aipLand: {
       display_name: "!LOC:AIP Land",
       percent_open_vehicle: 0.5,
       percent_open_bot: 0.5,
@@ -96,7 +96,7 @@ function qaipersonalities() {
       percent_naval: 0,
       percent_orbital: 0.05,
     },
-    "AIP Legonis Machina": {
+    aipLegonisMachina: {
       display_name: "!LOC:AIP Legonis Machina",
       fabber_to_factory_ratio_basic: 1.5,
       min_basic_fabbers: 1,
@@ -111,12 +111,12 @@ function qaipersonalities() {
       metal_demand_check: 0.75,
       energy_demand_check: 0.85,
     },
-    "AIP Low Tech": {
+    aipLowTech: {
       display_name: "!LOC:AIP Low Tech",
       adv_eco_mod: 3,
       adv_eco_mod_alone: 2,
     },
-    "AIP Naval": {
+    aipNaval: {
       display_name: "!LOC:AIP Naval",
       percent_open_vehicle: 0,
       percent_open_bot: 0,
@@ -129,7 +129,7 @@ function qaipersonalities() {
       percent_naval: 0.75,
       percent_orbital: 0.05,
     },
-    "AIP Orbital": {
+    aipOrbital: {
       display_name: "!LOC:AIP Orbital",
       percent_vehicle: 0.1125,
       percent_bot: 0.0625,
@@ -137,7 +137,10 @@ function qaipersonalities() {
       percent_naval: 0,
       percent_orbital: 0.775,
     },
-    "AIP Revenants": {
+    aipRandom: {
+      display_name: "!LOC:AIP Random",
+    },
+    aipRevenants: {
       display_name: "!LOC:AIP Revenants",
       fabber_to_factory_ratio_basic: 1.5,
       min_basic_fabbers: 1,
@@ -152,7 +155,7 @@ function qaipersonalities() {
       metal_demand_check: 0.75,
       energy_demand_check: 0.85,
     },
-    "AIP Rush": {
+    aipRush: {
       display_name: "!LOC:AIP Rush",
       percent_bot: 0.95,
       percent_vehicle: 0,
@@ -160,9 +163,9 @@ function qaipersonalities() {
       percent_naval: 0,
       percent_orbital: 0.05,
       neural_data_mod: 1.33,
-      adv_eco_mod: 3,
+      adv_eco_mod: 2,
     },
-    "AIP Swarm": {
+    aipSwarm: {
       display_name: "!LOC:AIP Swarm",
       percent_open_vehicle: 0,
       percent_open_bot: 1,
@@ -174,7 +177,7 @@ function qaipersonalities() {
       min_basic_fabbers: 3,
       min_advanced_fabbers: 1,
     },
-    "AIP Synchronous": {
+    aipSynchronous: {
       display_name: "!LOC:AIP Synchronous",
       fabber_to_factory_ratio_basic: 1.5,
       min_basic_fabbers: 1,
@@ -189,7 +192,7 @@ function qaipersonalities() {
       metal_demand_check: 0.75,
       energy_demand_check: 0.85,
     },
-    "AIP Tank": {
+    aipTank: {
       display_name: "!LOC:AIP Tank",
       percent_vehicle: 0.75,
       percent_bot: 0,
@@ -197,7 +200,7 @@ function qaipersonalities() {
       percent_naval: 0,
       percent_orbital: 0.05,
     },
-    "AIP Turtle": {
+    aipTurtle: {
       display_name: "!LOC:AIP Turtle",
       metal_drain_check: 0.71,
       energy_drain_check: 0.8,
@@ -218,18 +221,41 @@ function qaipersonalities() {
     personality,
     name
   ) {
-    var result = _.extend(_.clone(baseline), personality);
+    var result = _.assign(_.clone(baseline), personality);
     result["name"] = name;
     return result;
   });
 
-  _.extend(aiPersonalities, newPersonalities);
-
+  _.assign(aiPersonalities, newPersonalities);
   model.aiPersonalities.valueHasMutated();
+
+  model.startGame = (function () {
+    var cached_function = model.startGame;
+
+    return function () {
+      var absurdPersonalities = _.assign(
+        newPersonalities,
+        _.pick(ai_types(), "Absurd")
+      );
+
+      _.forEach(model.armies(), function (army) {
+        _.forEach(army.slots(), function (slot) {
+          if (slot.ai() === true && slot.aiPersonality() === "aipRandom")
+            slot.aiPersonality(
+              _.sample(_.keys(_.omit(absurdPersonalities, "aipRandom")))
+            );
+        });
+      });
+
+      var result = cached_function.apply(this, arguments);
+
+      return result;
+    };
+  })();
 }
 
 try {
-  qaipersonalities();
+  aipPersonalities();
 } catch (e) {
   console.error(e);
 }