+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v1.6.0 - 2020-10-05
+
+- Added AIP Random which will assign a random personality
+- Rush techs earlier
+
 ## v1.5.3 - 2020-07-12
 
 - All AIP- prefixes have had the hyphen removed to resolve a crash issue